+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to neopatch will be documented in this file.
 - In Touhou 18 and Touhou 20, holding Ctrl no longer fast-forwards replays while the game window is inactive.
 - Non-ASCII text in window titles and dialogs now correctly renders without mojibake even when using an incompatible system locale or ANSI code page.
 - neopatch no longer risks faulting from unmapped memory when the `dinput8.dll` file is loaded by a non-game executable.
+- D3D objects created by other software in the same process (e.g. overlays) are no longer accidentally affected by neopatch.
 
 ### Removed